chore: - redesign form layout in Bab raffle page
