+++ src/views/BabRaffle/index.tsx
@@ -9,8 +9,7 @@ import BabInfoCard from './components/BabInfoCard'
 import NFBGiveaway from './components/NFBGiveaway'
 import Services from 'views/Homepage/components/Services/Services'
 import NewToDeFi from './components/NewToDeFi'
-import MCFormContainer from 'components/Newsletter/MCFormContainer'
-// import Newsletter from 'components/Newsletter/Newsletter'
+import Newsletter from 'components/Newsletter'
 
 const Nft = () => {
   const { t } = useTranslation()
@@ -50,8 +49,7 @@ const Nft = () => {
           <Values />
         </SwiperProvider>
       </Flex>
-      <MCFormContainer />
-      {/* <Newsletter /> */}
+      <Newsletter />
     </Flex>
   )
 }
